Bump rig to 0.41 so reasoning shows on LocalAI (#745)
rig 0.39 only deserialized `delta.reasoning_content`. Providers that stream
reasoning as `delta.reasoning` — LocalAI, Cerebras — had it dropped by serde
while text and tool calls carried in the same delta worked fine, which is why
the panel stayed empty but everything else looked normal. rig 0.40 added the
fallback upstream; this takes us to 0.41.

The bump is wide. 0.41 splits the classic runtime into rig-agent behind an
`agent` feature, replaces Tool::definition with description/parameters across
all 37 tools, drops ToolDyn in favour of a concrete struct, and makes
Agent::preamble and Agent::model private. Tools move to the context-free
PortableTool contract, which is the honest fit — the loop has driven its own
dispatch since 4.5h-6 and never used rig's ToolContext. The erased-tool seam is
now dirge's own DynTool rather than rig's; it has broken on two consecutive rig
releases and does not need to be theirs. AnyAgentInner holds the completion
model directly instead of a rig Agent it only ever read the model back out of.

Capturing the reasoning then broke Cerebras, which rejects an echoed
reasoning_content on the next turn. The field is renamed to `reasoning` at the
wire boundary rather than dropped, so the model keeps its own reasoning in
context — the same thing @ai-sdk/cerebras does for opencode. Backends that want
reasoning_content are untouched: DeepSeek needs it on tool-call turns, and
llama.cpp/LocalAI chat templates read it back out of the assistant turn.

rmcp was on three versions at once — ours, one from rig's unused rmcp feature,
one from agent-client-protocol. That feature is gone, acp goes to 2.0 and rmcp
to 3.1, leaving one copy at latest. rusqlite, sysinfo, sha2, jsonschema, base64,
compact_str, http and notify-rust go current too; tree-sitter stays at 0.25
because 0.26 does not resolve against the grammar crates.

heal.rs listed a third repair it never implemented. The claim is gone, with a
note on why the port was skipped: opencode carries an equivalent for DeepSeek
only because its own transform lifts reasoning out of the message content, and
dirge replays the block in place.

+++ src/extras/session_db.rs
@@ -1876,14 +1876,17 @@ impl SessionDb {
             .map_err(|e| format!("Failed to prepare anchored view: {e}"))?;
 
         let messages: Vec<AnchorMessage> = stmt
-            .query_map(params![session_id, before + 1 + after, offset], |row| {
-                Ok(AnchorMessage {
-                    id: row.get(0)?,
-                    role: row.get(1)?,
-                    content: row.get(2)?,
-                    timestamp: row.get(3)?,
-                })
-            })
+            .query_map(
+                params![session_id, (before + 1 + after) as i64, offset],
+                |row| {
+                    Ok(AnchorMessage {
+                        id: row.get(0)?,
+                        role: row.get(1)?,
+                        content: row.get(2)?,
+                        timestamp: row.get(3)?,
+                    })
+                },
+            )
             .map_err(|e| format!("Failed to query anchored view: {e}"))?
             .filter_map(|r| r.ok())
             .collect();
